+++ packages/cli/src/lib/implementation/config-middleware.ts
@@ -1,5 +1,10 @@
 import { readCodePushupConfig } from '@code-pushup/core';
-import { CoreConfig } from '@code-pushup/models';
+import {
+  CoreConfig,
+  PERSIST_FILENAME,
+  PERSIST_FORMAT,
+  PERSIST_OUTPUT_DIR,
+} from '@code-pushup/models';
 import { GeneralCliOptions, OnlyPluginsOptions } from './model';
 import {
   filterCategoryByOnlyPluginsOption,
@@ -27,9 +32,22 @@ export async function configMiddleware<
         ...importedRc.upload,
         ...cliOptions.upload,
       },
+      // we can't use a async rc file as yargs does not support it. see: https://github.com/yargs/yargs/issues/2234
+      // therefore this can't live in option defaults as the order would be `config`->`provided options`->default
+      // so we have to manually implement the order
       persist: {
-        ...importedRc.persist,
-        ...cliOptions.persist,
+        outputDir:
+          cliOptions?.persist?.outputDir ||
+          importedRc?.persist?.outputDir ||
+          PERSIST_OUTPUT_DIR,
+        filename:
+          cliOptions?.persist?.filename ||
+          importedRc?.persist?.filename ||
+          PERSIST_FILENAME,
+        format:
+          cliOptions?.persist?.format ||
+          importedRc?.persist?.format ||
+          PERSIST_FORMAT,
       },
       plugins: filterPluginsByOnlyPluginsOption(importedRc.plugins, cliOptions),
       categories: filterCategoryByOnlyPluginsOption(