+++ packages/models/src/lib/core-config.ts
@@ -1,10 +1,6 @@
-import { Schema, z } from 'zod';
-import { CategoryConfig, categoryConfigSchema } from './category-config';
-import {
-  errorItems,
-  hasDuplicateStrings,
-  hasMissingStrings,
-} from './implementation/utils';
+import { z } from 'zod';
+import { categoriesSchema } from './category-config';
+import { errorItems, hasMissingStrings } from './implementation/utils';
 import { persistConfigSchema } from './persist-config';
 import { pluginConfigSchema } from './plugin-config';
 import { uploadConfigSchema } from './upload-config';
@@ -15,20 +11,10 @@ export const unrefinedCoreConfigSchema = z.object({
       'List of plugins to be used (official, community-provided, or custom)',
   }),
   /** portal configuration for persisting results */
-  persist: persistConfigSchema,
+  persist: persistConfigSchema.optional(),
   /** portal configuration for uploading results */
   upload: uploadConfigSchema.optional(),
-  categories: z
-    .array(categoryConfigSchema, {
-      description: 'Categorization of individual audits',
-    })
-    // categories slugs are unique
-    .refine(
-      categoryCfg => !getDuplicateSlugCategories(categoryCfg),
-      categoryCfg => ({
-        message: duplicateSlugCategoriesErrorMsg(categoryCfg),
-      }),
-    ),
+  categories: categoriesSchema,
 });
 
 export const coreConfigSchema = refineCoreConfig(unrefinedCoreConfigSchema);
@@ -38,17 +24,14 @@ export const coreConfigSchema = refineCoreConfig(unrefinedCoreConfigSchema);
  * workaround for zod issue: https://github.com/colinhacks/zod/issues/454
  *
  */
-export function refineCoreConfig(schema: Schema): Schema {
-  return (
-    schema
-      // categories point to existing audit or group refs
-      .refine(
-        coreCfg => !getMissingRefsForCategories(coreCfg),
-        coreCfg => ({
-          message: missingRefsForCategoriesErrorMsg(coreCfg),
-        }),
-      )
-  );
+export function refineCoreConfig(schema: typeof unrefinedCoreConfigSchema) {
+  // categories point to existing audit or group refs
+  return schema.refine(
+    coreCfg => !getMissingRefsForCategories(coreCfg),
+    coreCfg => ({
+      message: missingRefsForCategoriesErrorMsg(coreCfg),
+    }),
+  ) as unknown as typeof unrefinedCoreConfigSchema;
 }
 
 export type CoreConfig = z.infer<typeof unrefinedCoreConfigSchema>;
@@ -102,16 +85,4 @@ function getMissingRefsForCategories(coreCfg: CoreConfig) {
   }
 
   return missingRefs.length ? missingRefs : false;
-}
-
-// helper for validator: categories slugs are unique
-function duplicateSlugCategoriesErrorMsg(categories: CategoryConfig[]) {
-  const duplicateStringSlugs = getDuplicateSlugCategories(categories);
-  return `In the categories, the following slugs are duplicated: ${errorItems(
-    duplicateStringSlugs,
-  )}`;
-}
-
-function getDuplicateSlugCategories(categories: CategoryConfig[]) {
-  return hasDuplicateStrings(categories.map(({ slug }) => slug));
 }