+++ packages/models/src/lib/persist-config.ts
@@ -5,11 +5,9 @@ export const formatSchema = z.enum(['json', 'md']);
 export type Format = z.infer<typeof formatSchema>;
 
 export const persistConfigSchema = z.object({
-  outputDir: filePathSchema('Artifacts folder'),
-  filename: fileNameSchema('Artifacts file name (without extension)').default(
-    'report',
-  ),
-  format: z.array(formatSchema).default(['json']),
+  outputDir: filePathSchema('Artifacts folder').optional(),
+  filename: fileNameSchema('Artifacts file name (without extension)'),
+  format: z.array(formatSchema).default(['json']).optional(), // @TODO remove default or optional value and otherwise it will not set defaults.
 });
 
 export type PersistConfig = z.infer<typeof persistConfigSchema>;