+++ packages/utils/perf/crawl-file-system/optimized0.ts
@@ -1,0 +1,8 @@
+import { CrawlFileSystemOptions } from '../../src';
+
+export function crawlFileSystemOptimized0<T = string>(
+  options: CrawlFileSystemOptions<T>,
+): Promise<T[]> {
+  const { directory, pattern: needle = '@TODO' } = options;
+  return Promise.resolve([directory, needle] as T[]);
+}