+++ testing-utils/src/lib/fixtures/configs/code-pushup.invalid.config.ts
@@ -1,7 +1,5 @@
-import { join } from 'node:path';
-
 const config = {
-  persist: { outputDir: join('tmp', 'ts') },
+  persist: { outputDir: 'tmp' },
   upload: {
     organization: 'code-pushup',
     project: 'cli-ts',